requirements.txt
streamlit
agno
python-dotenv
groq
crawl4ai